requirements.txt
# SPDX-License-Identifier: MIT

Adafruit-Blinka
adafruit-circuitpython-minimqtt
paho-mqtt>=1.5.1